cleaning

+++ crypto/hblk_crypto.h
@@ -27,7 +27,7 @@
 typedef struct sig_s
 {
 	uint8_t *sig;
-	uint8_t len;
+	size_t len;
 } sig_t;
 
 /* Function declarations */
